requirements.txt
streamlit
requests
pandas
openpyxl
python-dotenv
